Add load-options! to load config from url option (-u)
New switches -u and --url <url> added to zprint.main

load-options! takes url, loads content as edn to pass to set-options!

Content is cached in new ~/.zprint/urlcache dir, respecting
Cache-Control max-age, the Expires header, or failing those the 5 min default.

If the url has bad content or is unavailable, an expired cache will be used but a warning will
be logging to System.err, else an Exception thrown.